Develop

+++ src/routes/payments.routes.js
@@ -1,8 +1,11 @@
 const express = require("express");
 const router = express.Router();
+const { saveAuditModel, decodedToken } = require("../utils");
+
 
 // Provider Model
 const Payments = require("../models/payment");
+const Sales = require("../models/sale");
 
 // GET all Payments
 router.get("/", async (req, res) => {
@@ -20,7 +23,7 @@ router.get("/:id", async (req, res) => {
 router.post("/", async (req, res) => {
   try {
     const { _id } = decodedToken(req);
-    const payment = new Payments(...req.body);
+    const payment = new Payments({ ...req.body, createdBy: _id });
     await payment.save();
 
     await saveAuditModel("Tipo de pago creado", _id);
@@ -38,7 +41,7 @@ router.put("/:id", async (req, res) => {
   try {
     const { _id } = decodedToken(req);
 
-    await Payments.findByIdAndUpdate(req.params.id, ...req.body);
+    await Payments.findByIdAndUpdate(req.params.id, { ...req.body });
 
     await saveAuditModel("Tipo de pago Actualizado", _id);
 
@@ -58,9 +61,23 @@ router.delete("/:id", async (req, res) => {
   try {
     const { _id } = decodedToken(req);
 
+    const sales = await Sales.find();
+
+    var ids = sales.reduce(
+      (acc, obj) => [...acc, obj.paymentType],
+      []
+    );
+
+    if (ids.some((id) => id == req.params.id)) {
+      return res.status(500).send({
+        success: false,
+        message: "Error",
+        error: "Hay formas de pago con este precio.",
+      });
+    }
     await Payments.findByIdAndRemove(req.params.id);
 
-    await saveAuditModel("Tipo de pago Eliminado", _id);
+    await saveAuditModel("Forma de pago Eliminada", _id);
 
     const payments = await Payments.find();
